Exceptions just use page URL not domain
Change to the config to allow users to just put the page url from the Backdrop root without the base_url (i.e. the domain) (and the parent folder root is not at domain root).  Made clear in the description.  This will simplify for users, but also avoid problems with people moving config from dev/staging/test to production.